ts fixes
