+++ packages/ts-moose-lib/package.json
@@ -38,7 +38,7 @@
     "@typescript-eslint/eslint-plugin": "^5.62.0",
     "@typescript-eslint/parser": "^5.62.0",
     "eslint": "^8.46.0",
-    "tsup": "^8.3.5"
+    "tsup": "^8.4.0"
   },
   "peerDependencies": {
     "ts-patch": "^3.3.0",
